Welcome to the function, team_5

team_5 returns a data frame of the geographic information of the polygons and the additional information such as name of the country, name of the territory/state.

team_5(file, tolerance)

Arguments

file
a character string of your shape file

tolerance
the value used for thinning the polygona; a single number of the tolerance value in the metric of the input object for an innerfunction called thinnedSpatialPolymaptools

Value

The return value is a data frame

Examples

ozfilepath <- system.file("extdata/gadm36_AUS_1.shp", package = "lab3Group9")
res <- team_5(ozfilepath,0.1)

#> Warning: package 'tidyverse' was built under R version 3.5.3

#> -- Attaching packages --------------------------------------------------------------------- tidyverse 1.2.1 --

#> v ggplot2 3.1.0       v purrr   0.3.2  
#> v tibble  2.1.1       v dplyr   0.8.0.1
#> v tidyr   0.8.3       v stringr 1.4.0  
#> v readr   1.3.1       v forcats 0.4.0  

#> Warning: package 'tibble' was built under R version 3.5.3

#> Warning: package 'purrr' was built under R version 3.5.3

#> -- Conflicts ------------------------------------------------------------------------ tidyverse_conflicts() --
#> x dplyr::filter()  masks stats::filter()
#> x purrr::is_null() masks testthat::is_null()
#> x dplyr::lag()     masks stats::lag()
#> x dplyr::matches() masks testthat::matches()

#> Linking to GEOS 3.6.1, GDAL 2.2.3, PROJ 4.9.3

head(res)

#>   group     long       lat order   country             territory.state
#> 1     1 123.5556 -12.53057     1 Australia Ashmore and Cartier Islands
#> 2     1 123.5554 -12.53213     2 Australia Ashmore and Cartier Islands
#> 3     1 123.5524 -12.53137     3 Australia Ashmore and Cartier Islands
#> 4     1 123.5536 -12.53039     4 Australia Ashmore and Cartier Islands
#> 5     1 123.5556 -12.53057     5 Australia Ashmore and Cartier Islands
#> 6     2 123.0178 -12.25917     1 Australia Ashmore and Cartier Islands

tail(res)

#>       group     long       lat order   country   territory.state
#> 53441  6707 124.2247 -15.94153     5 Australia Western Australia
#> 53442  6708 124.4609 -15.89704     1 Australia Western Australia
#> 53443  6708 124.4688 -15.93609     2 Australia Western Australia
#> 53444  6708 124.4490 -15.98464     3 Australia Western Australia
#> 53445  6708 124.4529 -15.90781     4 Australia Western Australia
#> 53446  6708 124.4609 -15.89704     5 Australia Western Australia

unique(res$territory.state)

#>  [1] Ashmore and Cartier Islands  Australian Capital Territory
#>  [3] Coral Sea Islands Territory  Jervis Bay Territory        
#>  [5] New South Wales              Northern Territory          
#>  [7] Queensland                   South Australia             
#>  [9] Tasmania                     Victoria                    
#> [11] Western Australia           
#> 11 Levels: Ashmore and Cartier Islands ... Western Australia